Player Movement Setup || Camera Setup
This commit introduces the player movement and camera setup. Player movement is going to be physics-based for reasons that will be clear later. The camera is set up in an isometric perspective.

Some issues with physics collisions are already evident, with a tower physics material taking care of many problems. However, it needs to be reminded that all towers must have the tower physics material attached to its collider.