Basic Mouse System & Simple Ability
This commit introduces a mouse setup for getting the mouse position in the world, using the same code from another project. Additionally, the logic for an ability was set up and tested. The player can perform a melee on an enemy now.

There is no clean way to get the mouse to my knowledge. Thus, a singleton mouse will exist. This mouse is responsible for getting the mouse position and any hit object. However, the code I copied had some leftover logic that must be moved to the proper script in due time. That script does not yet exist.