Beginning Tower System (#3)
* Update README.md

* Add Basic Input System Setup

This commit adds the beginnings of the input system. Reference the input architecture document for setup.

* Player Movement Setup || Camera Setup

This commit introduces the player movement and camera setup. Player movement is going to be physics-based for reasons that will be clear later. The camera is set up in an isometric perspective.

Some issues with physics collisions are already evident, with a tower physics material taking care of many problems. However, it needs to be reminded that all towers must have the tower physics material attached to its collider.

* Resolve Movement Bug

Fixed a bug with the code causing the player y velocity to exponentially grow every frame.

* Improve Movement Physics

This commit imrpoves the movement physics. Specifically, this commit aims to improve movement on sloped surfaces. Moving up surfaces is almost proper. However, If you hit the top of a sloped surface, then movement continues going upward, despite the fact the player should fall. Also, the player does not move down sloped surfaces properly. They fall down like it's stairs, hitting it constantly. These are minor bugs to work out in time.

* Basic Mouse System & Simple Ability

This commit introduces a mouse setup for getting the mouse position in the world, using the same code from another project. Additionally, the logic for an ability was set up and tested. The player can perform a melee on an enemy now.

There is no clean way to get the mouse to my knowledge. Thus, a singleton mouse will exist. This mouse is responsible for getting the mouse position and any hit object. However, the code I copied had some leftover logic that must be moved to the proper script in due time. That script does not yet exist.

* Begin Tower System

This commit starts building the tower system.

TODO: Add tower deployment criteria & placement logic, and work in the right branch!